add order delete but without ajax

+++ src/main/java/simonov/hotel/controllers/IndexController.java
@@ -84,7 +84,7 @@ public class IndexController {
     @ResponseBody
     Room updateRoom(@PathVariable int roomId,
                     @RequestParam String type,
-                    @RequestParam double price,
+                    @RequestParam int price,
                     @RequestParam String description,
                     @RequestParam int seats,
                     @RequestParam MultipartFile imageFile
@@ -119,7 +119,7 @@ public class IndexController {
     public String addRoom(@RequestParam String type,
                           @RequestParam int number,
                           @RequestParam String description,
-                          @RequestParam Double price,
+                          @RequestParam int price,
                           @RequestParam int seats,
                           @RequestParam int hotelId,
                           @RequestParam MultipartFile imageFile) {
